Moved method that fetches to chat
