+++ src/frontend/pages/chat/chat.js
@@ -37,6 +37,8 @@ export class Chat extends BaseComponent {
     messageToForward: "",
     activeDescription: "",
     scrolledToBottom: false,
+    hasMore: { type: Boolean },
+    hasFetchedNewMessages: { type: Boolean },
   };
 
   static get properties() {
@@ -56,6 +58,8 @@ export class Chat extends BaseComponent {
     super();
     this.messages = [];
     this.scrolledToBottom = false;
+    this.hasMore = true;
+    this.hasFetchedNewMessages = false;
 
     this.activeChatName =
       CookieService.getCookieByKey(CookieService.Keys.lastChat) || "";
@@ -91,10 +95,17 @@ export class Chat extends BaseComponent {
     ).reverse();
   }
 
-  async updated() {
-    await setTimeout(() => {
-      this.scrollToBottom();
-    }, 20);
+  async updated(changedProperties) {
+    if (changedProperties.has("messages")) {
+      if (changedProperties.has("hasFetchedNewMessages")) {
+        this.messagesListRef.value?.updateScrollPosition();
+        this.hasFetchedNewMessages = false;
+      } else {
+        await setTimeout(() => {
+          this.scrollToBottom();
+        }, 20);
+      }
+    }
   }
 
   static styles = css`
@@ -213,6 +224,7 @@ export class Chat extends BaseComponent {
                     .messages=${this.messages}
                     .activeChatName=${this.activeChatName}
                     .activeDescription=${this.activeDescription}
+                    .hasMore=${this.hasMore}
                     @il:message-forwarded=${this.openForwardMenu}
                     @il:went-to-chat=${this.wentToChatHandler}
                     @il:message-copied=${() =>
@@ -223,6 +235,7 @@ export class Chat extends BaseComponent {
                       )}
                     @il:message-edited=${this.editMessage}
                     @il:message-deleted=${this.askDeletionConfirmation}
+                    @il:update-next=${this.fetchNextMessages}
                   ></il-messages-list>
 
                   <il-modal
@@ -439,7 +452,7 @@ export class Chat extends BaseComponent {
   }
 
   async fetchMessages(e) {
-    this.messagesListRef.value?.resetHasMore();
+    this.hasMore = true;
 
     this.messages = (
       await MessagesService.getNextByRoomName(e.detail.conversation.roomName)
@@ -449,6 +462,34 @@ export class Chat extends BaseComponent {
     this.activeDescription = e.detail.conversation.description;
 
     this.inputControlsRef?.value?.focusEditor();
+  }
+
+  async fetchNextMessages(e) {
+    console.log(this.hasMore);
+    if (this.hasMore) {
+      let roomName = e?.detail?.conversation?.roomName;
+
+      if (!roomName) {
+        const cookie = CookieService.getCookie();
+        roomName = cookie.lastChat;
+      }
+
+      let before = null;
+      if (this.messages) {
+        before = this.messages[0].timestamp.replace(" ", "T");
+      }
+
+      let nextMessages = (
+        await MessagesService.getNextByRoomName(roomName, before)
+      ).reverse();
+
+      if (nextMessages.length === 0) {
+        this.hasMore = false;
+      } else {
+        this.messages = [...nextMessages, ...this.messages];
+        this.hasFetchedNewMessages = true;
+      }
+    }
   }
 
   createSocket() {
@@ -561,9 +602,7 @@ export class Chat extends BaseComponent {
 
       if (chatMessage.content !== null) {
         if (this.activeChatName == chatMessage.roomName) {
-          this.messages.push(chatMessage);
-          this.messagesListRef.value.requestUpdate();
-          this.requestUpdate();
+          this.messages = [...this.messages, chatMessage];
         }
 
         this.updateLastMessageInConversationList(chatMessage);
